More homogenous content column, keep sidebar for medium grid
The code for similar layouts is now more similar, this should help for further evolutions.

Ensure the 'row' div is used for every container. This helped fix the extra padding on /download.

Only when the main menu is collapsed will sidebars be pushed after the content.

Give more love to pages like /discover/