Improve OS X download page
Move the dmg description to the top, they apply to all versions.
Add icons so that download links stand out.
Move pertinent build instructions to the dedicated page.
Actually say that nightly has newer features.
Append "nightly" to the nightly link.
On homepage, switch Windows and Apple icons. They are now in the order of the download page.